injectfix 优化
