+++ Source/UnityProj/Assets/Helloworld/Calc.cs
@@ -45,7 +45,6 @@ public class Calculator
     private AllValueStruct astruct = new AllValueStruct{x=1};
 
     //修改成正确的逻辑后，打开如下注释，生成的补丁将修正该函数
-    [Patch]
     public int Add(int a, int b)
     {
         return 
@@ -55,7 +54,7 @@ public class Calculator
         TestEnum(TestEnumValue.t2) + TestEnum(default(TestEnumValue)) + TestEnum(thisEnum) + TestEnum(TestEnumValue.t1) + 
         TestRefInt(ref a) + a +
         TestNullable(null) + TestNullable(Vector3.left) + TestNullable(nullableV1) + TestNullable(nullableV2)
-        + TestNullable(nullableV3);
+        + TestNullable(nullableV3); 
         //return DoAdd(a, b); 
     }
 
@@ -95,14 +94,14 @@ public class Calculator
         return v.x * 10000;
     }
 
-    public int Sub(int a, int b)
-    {
-        return a / b;
-    }
-
     public int TestVector3(Vector3 v)
     {
         return 10 * (int)(v.x + v.y + v.z);
+    }
+
+    public int Sub(int a, int b)
+    {
+        return a / b;
     }
 
     public int Mult(int a, int b)
