+++ Source/UnityProj/Assets/IFix/IFixBindingCaller.New.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c8049ed4c8658d547a3082a48c3745e6
+guid: 3618f2a002a142d46aff20648fa8e4d8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
