+++ Source/UnityProj/Assets/IFix/TypeNameUtils.cs.meta
@@ -1,3 +1,11 @@
 fileFormatVersion: 2
-guid: 9a411fd5555e48788570e10b652f7d32
-timeCreated: 1711714781
+guid: 61c32234323c00b419afa29420efa8c9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
